working towards getting VMXON to work

+++ bfvmm/include/std/stdint.h
@@ -29,8 +29,8 @@ typedef unsigned char uint8_t;
 typedef short int int16_t;
 typedef unsigned short int uint16_t;
 
-typedef long int int32_t;
-typedef unsigned long int uint32_t;
+typedef int int32_t;
+typedef unsigned int uint32_t;
 
 typedef long long int int64_t;
 typedef unsigned long long int uint64_t;
